Added code for cats vs dogs
Addded code and model for acts vs dogs classification.
It includes split_data.py which splits the data into train val and test directories; and unzip.py which unzips the downloaded zip data. These need to be run individually.
plot_results.py plots loss and accuracy of a model.
model_gen loads a saved model for given parameters. If saved model not found then it trains a new model and saves it.
input_image.py asks the user to input an image for testing.